+++ src/main/java/nextstep/subway/line/LineController.java
@@ -1,7 +1,6 @@
 package nextstep.subway.line;
 
 import nextstep.subway.section.SectionRequest;
-import nextstep.subway.section.SectionService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -11,11 +10,9 @@ import java.util.List;
 @RestController
 public class LineController {
     private LineService lineService;
-    private SectionService sectionService;
 
-    public LineController(LineService lineService, SectionService sectionService) {
+    public LineController(LineService lineService) {
         this.lineService = lineService;
-        this.sectionService = sectionService;
     }
 
     @PostMapping("/lines")
@@ -52,13 +49,13 @@ public class LineController {
     @PostMapping("/lines/{id}/sections")
     public ResponseEntity<Void> addSection(@PathVariable Long id,
                                              @RequestBody SectionRequest sectionRequest) {
-        sectionService.addSection(id, sectionRequest);
+        lineService.addSection(id, sectionRequest);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/lines/{lineId}/sections")
     public ResponseEntity<LineResponse> deleteSection(@PathVariable Long lineId, @RequestParam Long stationId) {
-        sectionService.deleteSection(lineId, stationId);
+        lineService.deleteSection(lineId, stationId);
         return ResponseEntity.ok().build();
     }
