+++ src/main/java/nextstep/subway/section/SectionRequest.java
@@ -6,6 +6,12 @@ public class SectionRequest {
     private Long upStationId;
     private int distance;
 
+    public SectionRequest(Long upStationId, Long downStationId, int distance) {
+        this.upStationId = upStationId;
+        this.downStationId = downStationId;
+        this.distance = distance;
+    }
+
     public Long getDownStationId() {
         return downStationId;
     }
